Moved jax_np_interp into utils.py
