+++ diffstar/stars.py
@@ -9,6 +9,7 @@ import numpy as np
 from diffmah.individual_halo_assembly import _calc_halo_history
 from .quenching import quenching_function
 from .utils import _sigmoid, _inverse_sigmoid, _get_dt_array, tw_bin_jax_kern
+from .utils import jax_np_interp
 from .gas import _get_lagged_gas
 
 TODAY = 13.8
@@ -524,38 +525,6 @@ def _gas_conversion_kern(t_form, t_acc, dt, tau_dep):
         t_form,
     )
     return tri_kern
-
-
-@jjit
-def jax_np_interp(x, xt, yt, indx_hi):
-    """JAX-friendly implementation of np.interp.
-    Requires indx_hi to be precomputed, e.g., using np.searchsorted.
-
-    Parameters
-    ----------
-    x : ndarray of shape (n, )
-        Abscissa values in the interpolation
-    xt : ndarray of shape (k, )
-        Lookup table for the abscissa
-    yt : ndarray of shape (k, )
-        Lookup table for the ordinates
-
-    Returns
-    -------
-    y : ndarray of shape (n, )
-        Result of linear interpolation
-
-    """
-    indx_lo = indx_hi - 1
-    xt_lo = xt[indx_lo]
-    xt_hi = xt[indx_hi]
-    dx_tot = xt_hi - xt_lo
-    yt_lo = yt[indx_lo]
-    yt_hi = yt[indx_hi]
-    dy_tot = yt_hi - yt_lo
-    m = dy_tot / dx_tot
-    y = yt_lo + m * (x - xt_lo)
-    return y
 
 
 def fstar_tools(t_sim, fstar_tdelay=1.0):
